requirements.txt
# requirements.txt - 增强版，包含所有API支持

# 数据处理核心库
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# 数据获取 - 基础
yfinance>=0.2.0
pandas-datareader>=0.10.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# API客户端库
fredapi>=0.5.0
praw>=7.6.0

# 机器学习
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
tensorflow>=2.11.0
torch>=1.13.0

# 自然语言处理
transformers>=4.25.0
nltk>=3.8.0
vaderSentiment>=3.3.0
textblob>=0.17.0

# 统计分析
statsmodels>=0.13.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.12.0

# 工具库
python-dotenv>=0.19.0
tqdm>=4.64.0
joblib>=1.2.0
pathlib

# 日期时间处理
python-dateutil>=2.8.0

# JSON处理
ujson>=5.7.0

# 异步请求（可选，用于高频数据获取）
aiohttp>=3.8.0

# 数据验证
cerberus>=1.3.0

# 进度条
rich>=13.0.0